refacor(setup): main, App, router 역할 분리
